Feature/role based authorization

+++ BackendTascly/Services/ProjectService.cs
@@ -3,10 +3,14 @@ using BackendTascly.Repositories;
 
 namespace BackendTascly.Services
 {
-    public class ProjectService(IProjectsRepository projectsRepository) : IProjectService
+    public class ProjectService(IProjectsRepository projectsRepository, IUsersRepository usersRepository, IWorkspaceRepository workspaceRepository) : IProjectService
     {
-        public Task<bool> CreateProjectAsync(Project project, Guid userId, Guid workspaceId)
+        public async Task<bool> CreateProjectAsync(Project project, Guid userId, Guid workspaceId)
         {
+            // only workspace 'Admin' can create Project
+            var userRole = await workspaceRepository.GetWorkspaceUserRoleAsync(userId, workspaceId);
+            if (userRole is null || userRole.Name != "Admin") return false;
+
             project.WorkspaceId = workspaceId; // project must be created within a Workspace      
             project.OwnerId = userId; //assign owner to the project
 
@@ -23,17 +27,25 @@ namespace BackendTascly.Services
             project.TaskImportances.Add(new TaskImportance() { Name = "High" });
 
 
-            return projectsRepository.AddProjectAsync(project);
+            return await projectsRepository.AddProjectAsync(project);
         }
 
-        public Task<bool> DeleteProjectAsync(Guid projectId)
+        public async Task<bool> DeleteProjectAsync(Guid userId, Guid projectId)
         {
-            return projectsRepository.DeleteProjectAsync(projectId);
+            //get project to delete
+            var project = await projectsRepository.GetProjectById(projectId);
+            if (project is null) return false;
+
+            // only workspace 'Admin' can delete Project
+            var userRole = await workspaceRepository.GetWorkspaceUserRoleAsync(userId, project.WorkspaceId);
+            if (userRole is null || userRole.Name != "Admin") return false;
+
+            return await projectsRepository.DeleteProjectAsync(projectId);
         }
 
-        public Task<Project?> GetProjectByIdAsync(Guid projectId)
+        public async Task<Project?> GetProjectByIdAsync(Guid projectId)
         {
-            return projectsRepository.GetProjectById(projectId);
+            return await projectsRepository.GetProjectById(projectId);
         }
 
         public async Task<List<Project>> GetProjectsByWorkspaceId(Guid workspaceId)
