+++ BackendTascly/Services/TaskService.cs
@@ -4,7 +4,7 @@ using BackendTascly.Repositories;
 
 namespace BackendTascly.Services
 {
-    public class TaskService(ITaskRepository taskRepository) : ITaskService
+    public class TaskService(ITaskRepository taskRepository, IProjectsRepository projectsRepository, IWorkspaceRepository workspaceRepository) : ITaskService
     {
         public async Task<List<PTask>> GetTasksByProjectId(Guid projectId)
         {
@@ -22,6 +22,15 @@ namespace BackendTascly.Services
 
         public async Task<bool> CreateTaskAsync(PTask taskEntity, Guid userId, Guid projectId)
         {
+            //get project
+            var project = await projectsRepository.GetProjectById(projectId);
+            if (project is null) return false;
+
+            // only workspace 'Admin' or 'Full-access' can create a Task
+            var userRole = await workspaceRepository.GetWorkspaceUserRoleAsync(userId, project.WorkspaceId);
+            if (userRole is null || (userRole.Name != "Admin" && userRole.Name != "Full-access"))
+                return false;
+
             taskEntity.ProjectId = projectId; // Task must be created within a Project      
             taskEntity.AuthorId = userId; //assign task Author 
             taskEntity.CreationDate = DateTime.Now;
@@ -35,8 +44,21 @@ namespace BackendTascly.Services
             return await taskRepository.UpdateTaskAsync(taskEntity);
         }
 
-        public async Task<bool> DeleteTaskAsync(Guid taskId)
+        public async Task<bool> DeleteTaskAsync(Guid userId, Guid taskId)
         {
+            //get an existing task
+            var task = await taskRepository.GetTaskById(taskId);
+            if (task is null) return false;
+
+            //get project
+            var project = await projectsRepository.GetProjectById(task.ProjectId);
+            if (project is null) return false;
+
+            // only workspace 'Admin' or 'Full-access' can update a Task
+            var userRole = await workspaceRepository.GetWorkspaceUserRoleAsync(userId, project.WorkspaceId);
+            if (userRole is null || (userRole.Name != "Admin" && userRole.Name != "Full-access"))
+                return false;
+
             return await taskRepository.DeleteTaskAsync(taskId);
         }
     }
